requirements.txt
pycryptodome==3.9.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
PyQt5-stubs==5.14.2.2
